cleanup the puppeteer files
