+++ acceptance-tests/test/end-to-end/crossbrowser/supportedBrowsers.js
@@ -1,26 +1,35 @@
+/* eslint-disable object-curly-newline */
 const LATEST_MAC = 'macOS 10.15';
 const LATEST_WINDOWS = 'Windows 10';
 
 const supportedBrowsers = {
-  chrome: {
-    chrome_win_latest: {
-      browserName: 'chrome',
-      platformName: LATEST_WINDOWS,
-      browserVersion: 'latest',
-      'sauce:options': {
-        name: 'FeeAndPay: WIN_CHROME_LATEST',
-        extendedDebugging: true,
-        capturePerformance: true
-      }
-    },
-    chrome_mac_latest: {
-      browserName: 'chrome',
+  // This is safari browser
+  webkit: {
+    webkit_mac_latest: {
+      browserName: 'webkit',
       platformName: LATEST_MAC,
       browserVersion: 'latest',
       'sauce:options': {
-        name: 'FeeAndPay: MAC_CHROME_LATEST',
-        extendedDebugging: true,
-        capturePerformance: true
+        name: 'Mac_webkit_latest',
+        screenResolution: '1400x1050'
+      }
+    }
+  },
+  chromium: {
+    chromium_win_latest: {
+      browserName: 'chromium',
+      platformName: LATEST_WINDOWS,
+      browserVersion: 'latest',
+      'sauce:options': {
+        name: 'Win_chromium_latest'
+      }
+    },
+    chromium_mac_latest: {
+      browserName: 'chromium',
+      platformName: LATEST_MAC,
+      browserVersion: 'latest',
+      'sauce:options': {
+        name: 'Mac_chromium_latest'
       }
     }
   },
@@ -30,9 +39,15 @@ const supportedBrowsers = {
       platformName: LATEST_WINDOWS,
       browserVersion: 'latest',
       'sauce:options': {
-        name: 'FeeAndPay: WIN_FIREFOX_LATEST',
-        extendedDebugging: true,
-        capturePerformance: true
+        name: 'Win_Firefox_latest'
+      }
+    },
+    firefox_mac_latest: {
+      browserName: 'firefox',
+      platformName: LATEST_MAC,
+      browserVersion: 'latest',
+      'sauce:options': {
+        name: 'Mac_Firefox_latest'
       }
     }
   }
