--- bin/run-crossbrowser-tests.sh
@@ -1,17 +1,0 @@
-#!/bin/bash
-set -ex
-
-# Setup required environment variables. TEST_URL should be set by CNP
-export E2E_FRONTEND_URL=${TEST_URL}
-export E2E_PROXY_SERVER=${E2E_PROXY_SERVER:-"proxyout.reform.hmcts.net:8080"}
-export E2E_PROXY_BYPASS=${E2E_PROXY_BYPASS:-"*beta*LB.reform.hmcts.net"}
-export E2E_FRONTEND_NODE_ENV=${E2E_FRONTEND_NODE_ENV:-"production"}
-export E2E_WAIT_FOR_TIMEOUT_VALUE=${E2E_WAIT_FOR_TIMEOUT_VALUE:-15000}
-export E2E_WAIT_FOR_ACTION_VALUE=${E2E_WAIT_FOR_ACTION_VALUE:-250}
-export CODECEPT_PARAMS=${CODECEPT_PARAMS:-""}
-export E2E_TEST_FOR_CROSS_BROWSER='true'
-
-EXIT_STATUS=0
-yarn test:crossbrowser-e2e || EXIT_STATUS=$?
-echo EXIT_STATUS: $EXIT_STATUS
-exit $EXIT_STATUS
