--- bin/run-smoke-test.sh
@@ -1,4 +1,0 @@
-# Setup required environment variables. TEST_URL should be set by CNP
-export E2E_FRONTEND_URL=${TEST_URL}
-
-yarn test:smoketest
