+++ package.json
@@ -31,23 +31,17 @@
     "rename": "node ./scripts/rename.js",
     "replace": "node ./scripts/replace.js",
     "e2e": "ng e2e",
-    "test:smoketest": "codeceptjs run -c smoke-test/smoke-codecept.conf.js --reporter mochawesome --bail",
-    "test:acceptance": "cross-env LOG_LEVEL=OFF NODE_PATH=. ./node_modules/.bin/codeceptjs run-multiple parallel -c acceptance-tests/codecept.conf.js --grep @pipeline --verbose --reporter mocha-multi",
-    "test:acceptance-full": "cross-env LOG_LEVEL=OFF NODE_PATH=. ./node_modules/.bin/codeceptjs run-multiple parallel -c acceptance-tests/codecept.conf.js --grep @nightly --reporter mocha-multi",
-    "test:functional-local": "./bin/run-functional-tests-local.sh",
-    "test:acceptance-local": "cross-env LOG_LEVEL=OFF NODE_PATH=. ./node_modules/.bin/codeceptjs run -c acceptance-tests/codecept-local.conf.js --reporter mochawesome --bail",
     "sonar-scan": "sonar-scanner",
-    "test:crossbrowser-e2e": "cross-env LOG_LEVEL=OFF NODE_PATH=. ./node_modules/.bin/codeceptjs run-multiple --all --grep @crossbrowser -c acceptance-tests/saucelabs.conf.js --reporter mocha-multi",
     "test:mutation": "./bin/run-mutation-tests.sh",
     "test:stryker": "./node_modules/.bin/stryker run ",
     "test:smoke": "codeceptjs run --config smoke-test/smoke-codecept.conf.js",
     "test:functional": "npm-run-all -scl test:functional:pipeline test:functional:report",
     "test:fullfunctional": "npm-run-all -scl test:functional:nightly test:functional:report",
-    "test:functional:pipeline": "codeceptjs run-workers 6 parallel --config acceptance-tests/codecept-playwright.conf.js --grep @pipeline",
-    "test:functional:nightly": "codeceptjs run-workers 6 parallel --config acceptance-tests/codecept-playwright.conf.js --grep @nightly",
+    "test:functional:pipeline": "codeceptjs run-workers 6 parallel --config acceptance-tests/codecept.conf.js --grep @pipeline",
+    "test:functional:nightly": "codeceptjs run-workers 6 parallel --config acceptance-tests/codecept.conf.js --grep @nightly",
     "test:functional:report": "allure generate functional-output/functional/reports -c -o functional-output/functional/allure",
     "test:crossbrowser": "npm-run-all -scl test:crossbrowser:e2e test:crossbrowser:report",
-    "test:crossbrowser:e2e": "NODE_TLS_REJECT_UNAUTHORIZED=0 codeceptjs run-multiple --all --config acceptance-tests/saucelabs-playwright.conf.js --grep @crossbrowser",
+    "test:crossbrowser:e2e": "NODE_TLS_REJECT_UNAUTHORIZED=0 codeceptjs run-multiple --all --config acceptance-tests/saucelabs.conf.js --grep @crossbrowser",
     "test:crossbrowser:report": "allure generate functional-output/cross-browser/reports/* -c -o functional-output/cross-browser/allure"
   },
   "private": true,
